feat: add shared monitoring service
