+++ packages/admin-service/src/app.ts
@@ -8,6 +8,8 @@ import { createAdminRoutes } from './api/routes/admin.routes';
 import { MetricsService } from './services/metrics.service';
 import { ConfigService } from './services/config.service';
 import { ReportService } from './services/report.service';
+import { MonitoringService } from '@send/shared';
+import { logger } from '@shared/logger';
 
 const metricsService = new MetricsService();
 const configService = new ConfigService();
@@ -23,5 +25,17 @@ app.use(express.json());
 app.use(rateLimit('admin-service'));
 
 app.use('/api', createAdminRoutes(controller));
+
+const monitoringService = MonitoringService.getInstance();
+app.get('/metrics', async (_req, res) => {
+  try {
+    const metrics = await monitoringService.getMetrics();
+    res.set('Content-Type', 'text/plain');
+    res.send(metrics);
+  } catch (error) {
+    logger.error('Failed to get metrics:', error);
+    res.status(500).send('Failed to get metrics');
+  }
+});
 
 export default app;
