+++ packages/invoicing-service/src/app.ts
@@ -7,6 +7,8 @@ import { PrismaClient } from '@prisma/client';
 import { InvoiceService } from './services/invoice.service';
 import { InvoiceController } from './api/controllers/invoice.controller';
 import { createInvoiceRoutes } from './api/routes/invoice.routes';
+import { MonitoringService } from '@send/shared';
+import { logger } from '@shared/logger';
 
 const prisma = new PrismaClient();
 export const invoiceService = new InvoiceService(prisma);
@@ -21,6 +23,18 @@ app.use(express.json());
 app.use(rateLimit('invoicing-service'));
 
 app.use('/api/invoices', createInvoiceRoutes(invoiceController));
+
+const monitoringService = MonitoringService.getInstance();
+app.get('/metrics', async (_req, res) => {
+  try {
+    const metrics = await monitoringService.getMetrics();
+    res.set('Content-Type', 'text/plain');
+    res.send(metrics);
+  } catch (error) {
+    logger.error('Failed to get metrics:', error);
+    res.status(500).send('Failed to get metrics');
+  }
+});
 
 export { app, prisma };
 export default app;
