+++ packages/shared/src/index.ts
@@ -9,3 +9,4 @@ export { authenticate, requireRole } from './security/auth';
 export * from './responses';
 export { ValidationService, ValidationError } from './security/validation';
 export { createValidationErrorResponse } from './responses/error';
+export * from './monitoring';
