+++ packages/shared/src/monitoring/index.ts
@@ -1,0 +1,1 @@
+export * from './monitoring.service';
