+++ packages/student-service/src/app.ts
@@ -5,6 +5,8 @@ import compression from 'compression';
 import { securityHeaders, rateLimit } from '@send/shared/security/middleware';
 import studentRoutes from './api/routes/student.routes';
 import { errorHandler } from '@shared/errors';
+import { MonitoringService } from '@send/shared';
+import { logger } from '@shared/logger';
 
 const app = express();
 
@@ -18,6 +20,18 @@ app.use(rateLimit('student-service'));
 
 // Routes
 app.use('/api/students', studentRoutes);
+
+const monitoringService = MonitoringService.getInstance();
+app.get('/metrics', async (_req, res) => {
+  try {
+    const metrics = await monitoringService.getMetrics();
+    res.set('Content-Type', 'text/plain');
+    res.send(metrics);
+  } catch (error) {
+    logger.error('Failed to get metrics:', error);
+    res.status(500).send('Failed to get metrics');
+  }
+});
 
 // Error handling middleware
 app.use(errorHandler);
