+++ packages/user-service/package.json
@@ -29,8 +29,7 @@
     "express": "^4.18.2",
     "cors": "^2.8.5",
     "helmet": "^7.1.0",
-    "compression": "^1.7.4",
-    "express-prom-bundle": "8.0.0"
+    "compression": "^1.7.4"
   },
   "devDependencies": {
     "@types/express": "^4.17.21",
