+++ packages/user-service/src/app.ts
@@ -3,8 +3,7 @@ import cors from 'cors';
 import helmet from 'helmet';
 import compression from 'compression';
 import { securityHeaders, rateLimit } from '@send/shared/security/middleware';
-import promBundle from 'express-prom-bundle';
-import { MonitoringService } from './infra/monitoring/monitoring.service';
+import { MonitoringService } from '@send/shared';
 import { LoggingService } from 'shared/src/services/logging.service';
 import userRoutes from './api/routes/user.routes';
 import { errorHandler } from '@shared/errors';
@@ -14,12 +13,6 @@ const logger = new LoggingService('user-service');
 
 // Initialize monitoring
 const monitoringService = MonitoringService.getInstance();
-const metricsMiddleware = promBundle({
-    includeMethod: true,
-    includePath: true,
-    customLabels: { service: 'user-service' },
-    promRegistry: monitoringService.getRegistry(),
-});
 
 // Middleware
 app.use(securityHeaders);
@@ -32,9 +25,6 @@ app.use((req, res, next) => {
   logger.info('Incoming request', { method: req.method, url: req.url, ip: req.ip });
   next();
 });
-
-// Monitoring middleware
-app.use(metricsMiddleware);
 
 // Routes
 app.use('/api/users', userRoutes);
